refactor(web): move components
