+++ apps/web/src/components/transactions/TxDetails/TxData/Staking/StakingConfirmationTxDeposit.tsx
@@ -4,7 +4,7 @@ import FieldsGrid from '@/components/tx/FieldsGrid'
 import ConfirmationOrderHeader from '@/components/tx/ConfirmationOrder/ConfirmationOrderHeader'
 import { formatDurationFromMilliseconds, formatVisualAmount, maybePlural } from '@safe-global/utils/utils/formatters'
 import { formatCurrency } from '@safe-global/utils/utils/formatNumber'
-import StakingStatus from '../StakingStatus'
+import StakingStatus from './StakingStatus'
 import { InfoTooltip } from '@/components/common/InfoTooltip'
 import { BRAND_NAME } from '@/config/constants'
