+++ apps/web/src/components/transactions/TxDetails/TxData/Staking/StakingConfirmationTxWithdraw.tsx
