+++ apps/web/src/components/transactions/TxDetails/TxData/Staking/StakingStatus.tsx
